requirements.txt
Flask-PyMongo==2.3.0
Flask==1.1.1
twilio==6.33.1
reportlab==3.5.42
requests==2.22.0
python-dateutil==2.8.1
matplotlib==3.1.3
pandas==1.0.1
gunicorn==20.0.4
xmltodict==0.12.0